requirements.txt
tweepy>=4.14.0
anthropic>=0.40.0
groq>=0.9.0
feedparser>=6.0.0
python-dotenv>=1.0.0
